BSIS-937: Renamed PackTypeBasicViewModel
Renamed PackTypeBasicViewModel to PackTypeViewModel and old
PackTypeViewModel to PackTypeFullViewModel.

Note: there is no PackTypeFactory
