+++ src/viewmodel/PackTypeViewModel.java
@@ -4,56 +4,19 @@ import model.packtype.PackType;
 
 public class PackTypeViewModel {
 
-  private PackType packType;
+  private Long id;
+  private String packType;
 
   public PackTypeViewModel(PackType packType) {
-    this.packType = packType;
+    this.id = packType.getId();
+    this.packType = packType.getPackType();
   }
 
   public Long getId() {
-    return packType.getId();
+    return id;
   }
 
   public String getPackType() {
-    return packType.getPackType();
+    return packType;
   }
-
-  public ComponentTypeViewModel getComponentType() {
-    if (packType.getComponentType() != null) {
-      // FIXME: use factory
-      return new ComponentTypeViewModel(packType.getComponentType());
-    } else {
-      return null;
-    }
-  }
-
-  public Boolean getCanPool() {
-    return packType.getCanPool();
-  }
-
-  public Boolean getCanSplit() {
-    return packType.getCanSplit();
-  }
-
-  public Boolean getIsDeleted() {
-    return packType.getIsDeleted();
-  }
-
-  public Boolean getCountAsDonation() {
-    return packType.getCountAsDonation();
-  }
-
-  public Boolean getTestSampleProduced() {
-    return packType.getTestSampleProduced();
-  }
-
-  public Integer getPeriodBetweenDonations() {
-    return packType.getPeriodBetweenDonations();
-  }
-
-  @Override
-  public String toString() {
-    return packType.toString();
-  }
-
 }
